+++ paicorelib/framelib/__init__.py
@@ -1,3 +1,4 @@
+from .frame_gen import *
 from .frames import *
 
 # This library does not depend on any external libraries.